feat: Clean up codebase for production
- Remove all temporary test files and debug code
- Clean up imports and unused dependencies
- Remove commented code and debug prints
- Update README with current features and clean setup
- Add ENVIRONMENT_SETUP.md for configuration guide
- Remove monitoring directory (no longer needed)
- Organize project structure for production deployment
- Fix schema validation for estimated_duration field
- Ensure all LangGraph agents are working properly
- Ready for bootcamp submission and production use

+++ app/schemas.py
@@ -262,4 +262,4 @@ class AgentWorkflowResponse(BaseSchema):
     run_id: int
     status: str
     message: str
-    estimated_duration: Optional[int] = None  # seconds
+    estimated_duration: Optional[float] = None  # seconds
